AEA-3612 Copy getSecrets layer to seperate repo (#1)
# Description

:pencil: Summary:
The proposed changes to move the lambda layer for processing secrets to
a separate repository and publishing it as a release artifact for the
tracker API will have a significant business impact. This modification
ensures a more modular and scalable architecture, allowing for
independent updates and version control, thereby enhancing the
maintainability and flexibility of the system. It also facilitates the
seamless upgrade of the secret processing functionality for the tracker
API without disrupting other components, streamlining development and
deployment processes.

:ticket: Jira Reference:
<https://nhsd-jira.digital.nhs.uk/browse/AEA-3612>